requirements.txt
asgiref==3.5.2
Django==4.1.3
django-debug-toolbar==3.7.0
django-extensions==3.2.1
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6
